Fix getHasAccountInfo() check for Facebook to make sure Facebook favicon shows up in the accounts ribbon bar.
Tell the user when they are logged in to Facebook on the account page, give 
them an option to log out. If the user logs out, we set their session key valid 
flag to false and don't request any more updates for them. We can't actually 
make their session key invalid or log them out of Facebook.