Support resending an invitation to a group to any user or to an own contact. Don't make the user to whom an invitation is being sent a sender's contact.
Don't add people who invited you to a group to your contacts when you join.

Don't include other followers or people who are invited to follow in a drop-down 
of addable contacts for a follower.

Don't use the logic for whether the person can add members to the group when 
their status is INVITED, they can add self, but not other people.

Don't show invited members to a follower, don't show links to the invitation 
page on the invitation page.