Fix path to dumbhippo.tld in visit.jsp
Don't display presence icons for entity tags. It used to show a green presence icon when someone 
was online on Mugshot and had an AIM account listed. If we decide to have that again, we should
use an AIM presence key available from PersonView.

Make better use of the space in the framer:
  -- display four chat messages in the preview 
  --  make sure we fit three lines of description text