Fix two crashers:  - double-free of buddy->protocol in hippo-dbus-im.c  - NULL pointer deref in hippo-dbus-pidgin.c:pidgin_set_state These are both one-line fixes if we want to backport onto a branch.
Fix deadlock when a new client with the new bus name is run without --replace on the old client, and ends up calling ShowBrowser on itself

Fix a small memory leak in show_browser_blocking

Add a new hippo-dbus-helper feature for dealing with the comings and goings of a service (tracking NameOwnerChanged), and getting signals from the same service. Use this in hippo-dbus-pidgin.c to nuke a bunch of code before I cut-and-paste it into the new local presence engine.