Make mugshotEnabled nullable so that we have a third state when the user is indifferent to an account or have not specifically requested to not use the account on Mugshot.
Move all the instances of setting musghotEnabled to ExternalAcountSystem::setSentiment().

Set mugshotEnabled to be true on a single loved account of a type whenever one is created or whenever the previous one is deleted and another one exists. Set mugshoEnabled to be null when the sentiment changes to being indifferent.

Also move the check for making sure that we only have a single hated account for a type and no other loved accounts can exist when there is a hated one to ExternalAcountSystem::setSentiment().